% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assortative_mating_functions.R
\name{assort_mating}
\alias{assort_mating}
\title{Generate the pairs selected by mothers}
\usage{
assort_mating(popsize, female_phenotype, male_phenotype, corr_coeff)
}
\arguments{
\item{female_phenotype}{A numeric vector giving the phenotypes of all females}

\item{male_phenotype}{A numeric phenotype giving the phenotypes of all males}

\item{corr_coeff}{The correlation coefficient between phenotypes for males and females}

\item{pop_size}{The number of pairs to generate.}
}
\value{
the indices of the returned mothers and fathers
}
\description{
Generate the pairs selected by mothers
}
