% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{calculate_phenotype}
\alias{calculate_phenotype}
\title{Get the phenotype and the additive genetic part of the trait}
\usage{
calculate_phenotype(eff, genotype)
}
\arguments{
\item{genotype}{The genotypes with individuals in columns and sites in rows.}

\item{effects}{The effect size object}
}
\value{
A data frame with the phenotype and additive genetic variance.
}
\description{
Gets the phenotype for each individual for the a genetic
effect which comprises the additive genetic effect, and the
environmental effect.e
}
