% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assortative_mating_functions.R
\name{generate_effects}
\alias{generate_effects}
\title{Get the genotype effects for each locus for two traits.}
\usage{
generate_effects(loci, allele_freq, env_var)
}
\arguments{
\item{loci}{A vector with the loci affecting the trait}

\item{allele_freq}{The allele frequency of sampled loci.
A vector at least as long as max loci specified.}
}
\value{
A list with elements of the effect and the genetic standard deviation on the trait of class effects
}
\description{
The genotype effects consist of the effect for each locus, which is drawn from a Normal(0,1)
distribution for each locus and the  genetic standard deviation of that trait which is the
standard deviation for each trait normalised by the allele frequency.
}
