% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_sims.R
\name{generate_QT_assortative_mating}
\alias{generate_QT_assortative_mating}
\title{Generate father mother pairs}
\usage{
generate_QT_assortative_mating(
  population_size,
  loci_A,
  loci_B,
  environ_var,
  corr_coeff = 0,
  n_loci,
  allele_freq = 0.5
)
}
\arguments{
\item{population_size}{The Size of the populations to simulate.  Equal numbers of males and
females of this size.}

\item{loci_A}{loci affecting trait A}

\item{loci_B}{loci affecting trait B}

\item{environ_var}{environmental variance}

\item{corr_coeff}{Correlation coefficient between phenotype in males and females}

\item{n_loci}{The number of loci to simulate.  Not always needed.}

\item{allele_freq}{The allele frequency of sampled loci.
Either a single value or a vector at least as long as max loci specified.}
}
\description{
This is really just a test that the simulations and the
correlation coefficients work.
}
\examples{
g1 <- generate_QT_assortative_mating(population_size = 100000, loci_A = 1:100, loci_B = 101:200,
                                     environ_var = 0.3, corr_coeff = 0.3)
cor(g1$mother_phenotype$pheno, g1$father_phenotype$pheno)#
plot(g1$mother_phenotype$pheno, g1$father_phenotype$pheno)
# Simulate for loci with rarer variants
g2 <- generate_QT_assortative_mating(population_size = 100000, loci_A = 1:100, loci_B = 101:200,
                         environ_var = 0.3, corr_coeff = 0.3, allele_freq = runif(200, 0.01, 0.05))
cor(g2$mother_phenotype$pheno, g2$father_phenotype$pheno)
plot(g2$mother_phenotype$pheno, g2$father_phenotype$pheno)

}
