% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{generation}
\alias{generation}
\title{Get the next generation of Genotypes}
\usage{
generation(g, effects_A, effects_B, corr_coeff)
}
\arguments{
\item{g}{a list containing male_genotype and female_genotype}

\item{effects_A}{the effects for trait A}

\item{effects_B}{the effects for trait B}

\item{corr_coeff}{Correlation coefficient between traits for assortative mating}
}
\description{
Get the next generation of Genotypes
}
\examples{


}
