% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{heritability}
\alias{heritability}
\title{Find the heritability for a trait}
\usage{
heritability(phen)
}
\arguments{
\item{phen}{The Phenotype.  An object of type phenotype}
}
\value{
the heritability
}
\description{
Find the heritability for a trait
}
\examples{
allele_freq <- runif(2000, 0.01, 0.2)
heritability(generate_phenotype(eff=generate_effects(1:1000, allele_freq, 1), sample_initial_genotypes(3000, allele_freq)))
}
