% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{init}
\alias{init}
\title{Initialise the genotypes and effects and get a set of male and female genotypes}
\usage{
init(
  population_size,
  loci_A,
  loci_B,
  corr_coeff,
  environ_var = c(1, 1),
  n_loci,
  allele_freq = 0.5
)
}
\arguments{
\item{population_size}{The size of the population to simulate}

\item{loci_A}{Loci affecting trait A only}

\item{loci_B}{Loci affecting trait B only}

\item{loci_AB}{Loci affecting traits A and B}

\item{corr_coefficient}{Genetic correlation between traits A and B based on shared loci}
}
\description{
Initialise the genotypes and effects and get a set of male and female genotypes
}
