% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assortative_mating_functions.R
\name{make_a_gamete}
\alias{make_a_gamete}
\title{Generate a gamete based on the parental genotype}
\usage{
make_a_gamete(parental_genotype)
}
\arguments{
\item{parental_genotype}{The genotype (as allele count) of the parent}
}
\value{
A vector with the haplotype (as allele count)
}
\description{
Generate a gamete based on the parental genotype
}
