% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{next_generation}
\alias{next_generation}
\title{Get the next generation of genotypes}
\usage{
next_generation(mother_genotypes, father_genotypes)
}
\arguments{
\item{mother_genotypes}{loci affecting trait A}

\item{father_genotypes}{loci affecting trait B}
}
\description{
Get the next generation of genotypes
}
\examples{


}
