% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{plot.phenotype}
\alias{plot.phenotype}
\title{Plot a phenotype}
\usage{
\method{plot}{phenotype}(phen)
}
\arguments{
\item{phen}{the phenotype}
}
\value{
NULL, run for side effects
}
\description{
Plot a phenotype
}
\examples{
allele_freq <- runif(2000, 0.01, 0.2)
plot(generate_phenotype(eff=generate_effects(1:1000, allele_freq, .2), sample_initial_genotypes(3000, allele_freq)))
}
