% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/am_sims.R
\name{QT_assortative_mating_report}
\alias{QT_assortative_mating_report}
\title{Provide a quick report on the simulation of parental pairs and a child.}
\usage{
QT_assortative_mating_report(
  population_size,
  loci_A,
  loci_B,
  environ_var,
  corr_coeff = 0,
  n_loci,
  allele_freq = 0.5,
  plots = FALSE,
  write_plots = FALSE
)
}
\arguments{
\item{population_size}{The Size of the populations to simulate.  Equal numbers of males and
females of this size.}

\item{loci_A}{loci affecting trait A}

\item{loci_B}{loci affecting trait B}

\item{environ_var}{environmental variance}

\item{corr_coeff}{Correlation coefficient between phenotype in males and females}

\item{n_loci}{The number of loci to simulate.  Not always needed.}

\item{allele_freq}{The allele frequency of sampled loci.
Either a single value or a vector at least as long as max loci specified.}
}
\description{
This report is only for a single generation
}
