% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{sample_initial_genotypes}
\alias{sample_initial_genotypes}
\title{Sample the starting genotypes}
\usage{
sample_initial_genotypes(num_inds, allele_freq)
}
\arguments{
\item{num_inds}{Number of individuals}

\item{allele_freq}{Allele frequency for each locus}
}
\value{
A matrix of allele counts, individuals in columns, loci in rows.
}
\description{
Sample the starting genotypes
}
\examples{
a <- sample_initial_genotypes(100, rep(0.5, 20))
allele_freq <- runif(20, 0.01, 0.99)
b <- sample_initial_genotypes(100, allele_freq))
d <- rowMeans(b)/2
plot(allele_freq, d)

}
