% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assortative_mating_functions.R
\name{sexual_reproduction}
\alias{sexual_reproduction}
\title{Get the next generation of Genotypes}
\usage{
sexual_reproduction(nchildren, g, effects_A, effects_B, corr_coeff)
}
\arguments{
\item{nchildren}{The number of children to generate}

\item{g}{a list containing male_genotype and female_genotype}

\item{effects_A}{the effects for trait A}

\item{effects_B}{the effects for trait B}

\item{corr_coeff}{Correlation coefficient between traits for assortative mating}
}
\description{
Get the next generation of Genotypes
}
\examples{


}
