% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortative_mating_functions.R
\name{generate_bio_pleiotrophy_effects}
\alias{generate_bio_pleiotrophy_effects}
\title{Get correlated effects for a pair of traits.}
\usage{
generate_bio_pleiotrophy_effects(loci, allele_freq, corr_coeff, env_var)
}
\arguments{
\item{loci}{A vector with the loci affecting the traits}

\item{allele_freq}{The allele frequency of sampled loci.
A vector at least as long as max loci specified.}

\item{corr_coeff}{The correlation coefficient between traits}

\item{env_var}{The environmental variance.  Either one value shared by both traits or
a vector of length 2.}
}
\value{
A a list with effects_A and effects_B
}
\description{
The genotype effects consist of the effect for each locus, which is drawn from a Normal(0,1)
distribution for each locus and the genetic standard deviation of that trait which is the
standard deviation for each trait normalised by the allele frequency.
}
